arm64: dts: qcom: x1e80100-denali: Use samsung,atna30dw01 for panel
The Microsoft Surface Pro 11 (Snapdragon) ships with two panel options;
an LCD model and an OLED model. The OLED model is using the SDC 4195,
aka. Samsung ATNA30DW01-1.

Whilst panel-edp works fine with the OLED version of this device as-is,
a dmesg warning is logged due to it being unknown.

The samsung,atna33xc20 driver is probably a more appropriate choice of
panel driver for this machine as it gets timings more correct for
Samsung OLED panels.

Change the devicetree entry to use this compatible driver instead.

Signed-off-by: Dale Whinham <daleyo@gmail.com>